Paging chunks is working.  Needs optimized heavily.

+++ src/org/blockworld/level/neighbor/BottomNeighbor.java
@@ -18,7 +18,7 @@ public class BottomNeighbor implements Neighbor {
 	 * @see org.blockworld.level.neighbor.Neighbor#getGeometry(int, int, int, int)
 	 */
 	@Override
-	public Geometry getGeometry(int x, int y, int z, final Vector2f chunkPosition, int levelSize, int bias) {
+	public Geometry getGeometry(float x, float y, float z, final Vector2f chunkPosition, int levelSize, int bias) {
 		final Quad q = new Quad(1.0f, 1.0f);
 		final Geometry g = new Geometry("Box" + x + y + z, q);
 		Vector2f pos = chunkPosition.mult(levelSize);
