+++ src/org/blockworld/level/neighbor/TopNeighbor.java
@@ -15,7 +15,7 @@ import com.jme3.scene.shape.Quad;
 public class TopNeighbor implements Neighbor {
 
 	@Override
-	public Geometry getGeometry(int x, int y, int z, final Vector2f chunkPosition, int levelSize, int bias) {
+	public Geometry getGeometry(float x, float y, float z, final Vector2f chunkPosition, int levelSize, int bias) {
 		final Quad q = new Quad(1.0f, 1.0f);
 		final Geometry g = new Geometry("Box" + x + y + z, q);
 		Vector2f pos = chunkPosition.mult(levelSize);
